+++ packages/plugin/src/estree-parser/estree-ast.ts
@@ -1,10 +1,10 @@
-import { ASTNode, TypeInfo, TypeNode, ValueNode } from 'graphql';
-import { BaseNode } from 'estree';
+import type { ASTNode, TypeInfo, TypeNode } from 'graphql';
+import type { SourceLocation, Comment } from 'estree';
+import type { AST } from 'eslint';
 
-type SafeGraphQLType<T extends ASTNode | ValueNode> = Omit<
-  T extends { readonly type: TypeNode } ? Omit<T, 'type'> & { readonly gqlType: TypeNode } : T,
-  'loc'
->;
+type SafeGraphQLType<T extends ASTNode> = T extends { type: TypeNode }
+  ? Omit<T, 'loc' | 'type'> & { gqlType: TypeNode }
+  : Omit<T, 'loc'>;
 
 export type TypeInformation = {
   argument: ReturnType<TypeInfo['getArgument']>;
@@ -18,23 +18,21 @@ export type TypeInformation = {
   gqlType: ReturnType<TypeInfo['getType']>;
 };
 
-type SingleESTreeNode<T extends ASTNode | ValueNode, WithTypeInfo extends boolean> = SafeGraphQLType<T> &
-  Pick<BaseNode, 'leadingComments' | 'loc' | 'range'> & {
-    type: T['kind'];
-    // eslint-disable-next-line @typescript-eslint/ban-types -- Record<string, never> don't work
-    typeInfo: () => WithTypeInfo extends true ? TypeInformation : {};
-    rawNode: () => T;
-  };
+type Node<T extends ASTNode, WithTypeInfo extends boolean> = SafeGraphQLType<T> & {
+  type: T['kind'];
+  loc: SourceLocation;
+  range: AST.Range;
+  leadingComments: Comment[];
+  typeInfo: () => WithTypeInfo extends true ? TypeInformation : Record<string, never>;
+  rawNode: () => T;
+};
 
-// eslint-disable-next-line @typescript-eslint/ban-types -- only empty object makes the tooltips expand without recursion
-export type GraphQLESTreeNode<T, WithTypeInfo extends boolean = false> = {} & (T extends ASTNode | ValueNode
+export type GraphQLESTreeNode<T, W extends boolean = false> = T extends ASTNode
   ? {
-      [K in keyof SingleESTreeNode<T, WithTypeInfo>]: SingleESTreeNode<T, WithTypeInfo>[K] extends ReadonlyArray<
-        infer Nested
-      >
-        ? GraphQLESTreeNode<Nested, WithTypeInfo>[]
-        : SingleESTreeNode<T, WithTypeInfo>[K] extends ASTNode
-        ? GraphQLESTreeNode<SingleESTreeNode<T, WithTypeInfo>[K], WithTypeInfo>
-        : SingleESTreeNode<T, WithTypeInfo>[K];
+      [K in keyof Node<T, W>]: Node<T, W>[K] extends ReadonlyArray<infer Nested>
+        ? GraphQLESTreeNode<Nested, W>[]
+        : Node<T, W>[K] extends ASTNode
+        ? GraphQLESTreeNode<Node<T, W>[K], W>
+        : Node<T, W>[K];
     }
-  : T);
+  : T;
