+++ packages/plugin/src/estree-parser/utils.ts
@@ -9,7 +9,7 @@ import {
   Source,
   Lexer,
 } from 'graphql';
-import type { Comment } from 'estree';
+import type { Comment, SourceLocation } from 'estree';
 import type { AST } from 'eslint';
 import { valueFromASTUntyped } from 'graphql/utilities/valueFromASTUntyped';
 
@@ -113,7 +113,7 @@ export function extractComments(loc: Location): Comment[] {
   const comments: Comment[] = [];
   let token = loc.startToken;
 
-  while (token !== null) {
+  while (token) {
     if (token.kind === TokenKind.COMMENT) {
       const comment = convertToken(
         token,
@@ -125,4 +125,30 @@ export function extractComments(loc: Location): Comment[] {
     token = token.next;
   }
   return comments;
+}
+
+export function convertLocation(location: Location): SourceLocation {
+  const { startToken, endToken, source, start, end } = location;
+  /*
+   * ESLint has 0-based column number
+   * https://eslint.org/docs/developer-guide/working-with-rules#contextreport
+   */
+  const loc = {
+    start: {
+      /*
+       * Kind.Document has startToken: { line: 0, column: 0 }, we set line as 1 and column as 0
+       */
+      line: startToken.line === 0 ? 1 : startToken.line,
+      column: startToken.column === 0 ? 0 : startToken.column - 1,
+    },
+    end: {
+      line: endToken.line,
+      column: endToken.column - 1,
+    },
+    source: source.body,
+  };
+  if (loc.start.column === loc.end.column) {
+    loc.end.column += end - start;
+  }
+  return loc;
 }
