+++ packages/plugin/src/index.ts
@@ -1,6 +1,6 @@
 export { rules } from './rules';
 export { processors } from './processors';
-export * from './parser';
+export { parseForESLint } from './parser';
 export * from './types';
 export * from './testkit';
 export { requireGraphQLSchemaFromContext, requireSiblingsOperations } from './utils';
