+++ packages/plugin/src/rules/no-anonymous-operations.ts
@@ -44,7 +44,7 @@ const rule: GraphQLESLintRule = {
       'OperationDefinition[name=undefined]'(node: GraphQLESTreeNode<OperationDefinitionNode>) {
         const [firstSelection] = node.selectionSet.selections;
         const suggestedName =
-          firstSelection.type === Kind.FIELD ? (firstSelection.alias || firstSelection.name).value : node.operation;
+          firstSelection.kind === Kind.FIELD ? (firstSelection.alias || firstSelection.name).value : node.operation;
 
         context.report({
           loc: getLocation(node.loc.start, node.operation),
